HomeWork_12/pyproject.toml
[tool.poetry]
name = "homework-12"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "homework_12"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
sqlalchemy = "^2.0.8"
psycopg2 = "^2.9.5"
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
alembic = "^1.10.2"
pydantic = {extras = ["email"], version = "^1.10.7"}
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
